Move test_st_time_input test to time_widgets_test.py
