+++ lib/tests/streamlit/streamlit_test.py
@@ -14,7 +14,6 @@
 
 """Streamlit Unit test."""
 
-import datetime
 import os
 import re
 import subprocess
@@ -176,40 +175,6 @@ class StreamlitTest(unittest.TestCase):
 
 class StreamlitAPITest(DeltaGeneratorTestCase):
     """Test Public Streamlit Public APIs."""
-
-    def test_st_time_input(self):
-        """Test st.time_input."""
-        value = datetime.time(8, 45)
-        st.time_input("Set an alarm for", value)
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.time_input.default, "08:45")
-        self.assertEqual(el.time_input.step, datetime.timedelta(minutes=15).seconds)
-
-    def test_st_time_input_with_step(self):
-        """Test st.time_input with step."""
-        value = datetime.time(9, 00)
-        st.time_input("Set an alarm for", value, step=datetime.timedelta(minutes=5))
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.time_input.default, "09:00")
-        self.assertEqual(el.time_input.step, datetime.timedelta(minutes=5).seconds)
-
-    def test_st_time_input_exceptions(self):
-        """Test st.time_input exceptions."""
-        value = datetime.time(9, 00)
-        with self.assertRaises(StreamlitAPIException):
-            st.time_input("Set an alarm for", value, step=True)
-        with self.assertRaises(StreamlitAPIException):
-            st.time_input("Set an alarm for", value, step=(90, 0))
-        with self.assertRaises(StreamlitAPIException):
-            st.time_input("Set an alarm for", value, step=1)
-        with self.assertRaises(StreamlitAPIException):
-            st.time_input("Set an alarm for", value, step=59)
-        with self.assertRaises(StreamlitAPIException):
-            st.time_input("Set an alarm for", value, step=datetime.timedelta(hours=24))
-        with self.assertRaises(StreamlitAPIException):
-            st.time_input("Set an alarm for", value, step=datetime.timedelta(days=1))
 
     def test_st_legacy_vega_lite_chart(self):
         """Test st._legacy_vega_lite_chart."""
